+++ src/react/components/dateinput/private/XUIDateInputItem.js
@@ -332,6 +332,7 @@ class XUIDateInputItem extends Component {
             id={this.suggestedDatesFooterId}
             leftElement={<XUIIcon color="black" icon={selectDateIcon || dateStartIcon} isBoxed />}
             onClick={this.showDatepickerPanel}
+            qaHook={qaHook && `${qaHook}-dateinputitem-suggesteddates--footer`}
           >
             {selectDateLabel}
           </XUIPickitem>
@@ -400,6 +401,7 @@ class XUIDateInputItem extends Component {
         restrictedToViewPort={false}
         trigger={trigger}
         triggerClickAction="none"
+        useNewFocusBehaviour
       />
     );
   }