+++ src/react/filepreview.ts
@@ -1,0 +1,5 @@
+import XUIFilePreview from './components/filepreview/XUIFilePreview';
+import XUIFilePreviewFooter from './components/filepreview/XUIFilePreviewFooter';
+import XUIFilePreviewHeader from './components/filepreview/XUIFilePreviewHeader';
+
+export { XUIFilePreview as default, XUIFilePreview, XUIFilePreviewFooter, XUIFilePreviewHeader };